fixed args and added README.md

+++ egs/demask/run.sh
@@ -6,10 +6,10 @@ set -o pipefail
 
 # Main storage directory. You'll need disk space to dump the WHAM mixtures and the wsj0 wav
 # files if you start from sphere files.
-storage_dir=
+storage_dir=./datasets
 
-librispeech_dir=/media/sam/Data/LibriSpeech/
-rir_dir=$storage_dir/rir_data
+librispeech_dir=/media/sam/Data/LibriSpeech #$storage_dir/LibriSpeech
+rir_dir=/media/sam/Data/dcase/rir_data #$storage_dir/rir_data
 # After running the recipe a first time, you can run it from stage 3 directly to train new models.
 
 # Path to the python you'll use for the experiment. Defaults to the current python
@@ -41,32 +41,45 @@ train_dir=$dumpdir/tr
 valid_dir=$dumpdir/cv
 test_dir=$dumpdir/tt
 
+
+# Check if sox is installed
+if ! [[ -x "$(command -v sox)" ]] ; then
+  echo "This recipe requires SoX, Install sox with `conda install -c conda-forge sox`. Exiting."
+  exit 1
+fi
+
+# Install pysndfx if not instaled
+if not python -c "import pysndfx" &> /dev/null; then
+    echo 'This recipe requires pysndfx. Installing requirements.'
+    $python_path -m pip install -r requirements.txt
+fi
+
 if [[ $stage -le  0 ]]; then
   echo "Stage 0: Downloading required Datasets"
   echo "Downloading LibriSpeech/train-clean-360 into $librispeech_dir"
 	# If downloading stalls for more than 20s, relaunch from previous state.
-	wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/train-clean-360.tar.gz -P $librispeech_dir
-	tar -xzf $storage_dir/train-clean-360.tar.gz -C $librispeech_dir
-	rm -rf $storage_dir/train-clean-360.tar.gz
 
-	wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/dev-clean.tar.gz -P $librispeech_dir
-	tar -xzf $storage_dir/dev-clean.tar.gz -C $librispeech_dir
+	wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/dev-clean.tar.gz -P $storage_dir
+	tar -xzf $storage_dir/dev-clean.tar.gz -C $storage_dir
 	rm -rf $storage_dir/dev-clean.tar.gz
 
+	wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/dev-clean.tar.gz -P $storage_dir
+	tar -xzf $storage_dir/dev-clean.tar.gz -C $storage_dir
+	rm -rf $storage_dir/dev-clean.tar.gz
 
-  wget -c --tries=0 --read-timeout=20 https://zenodo.org/record/3743844/files/FUSS_rir_data.tar.gz $rir_dir
-	tar -xzf $rir_dir/FUSS_rir_data.tar.gz -C $rir_dir
-	rm -rf $rir_dir/FUSS_rir_data.tar.gz
+  wget -c --tries=0 --read-timeout=20 https://zenodo.org/record/3743844/files/FUSS_rir_data.tar.gz $storage_dir
+	tar -xzf $storage_dir/FUSS_rir_data.tar.gz -C $rir_dir
+	rm -rf $storage_dir/FUSS_rir_data.tar.gz
 
 fi
 
 if [[ $stage -le  1 ]]; then
 	echo "Stage 1: parsing the datasets"
 	for librispeech_split in train-clean-360 dev-clean ; do
-    python local/parse_data.py --input_dir $librispeech_dir/$librispeech_split --output_json $dumpdir/${librispeech_split}.json --regex **/*.flac
+    python local/parse_data.py --input_dir $librispeech_dir/$librispeech_split --output_json $dumpdir/clean/${librispeech_split}.json --regex **/*.flac
   done
   for rir_split in train validation ; do
-    python local/parse_data.py --input_dir $rir_dir/$rir_split --output_json $dumpdir/${rir_split}.json --regex **/*.wav
+    python local/parse_data.py --input_dir $rir_dir/$rir_split --output_json $dumpdir/rirs/${rir_split}.json --regex **/*.wav
   done
 fi
 
@@ -83,11 +96,11 @@ if [[ $stage -le 2 ]]; then
   echo "Stage 3: Training"
   mkdir -p logs
   CUDA_VISIBLE_DEVICES=$id $python_path train.py \
-		--clean_speech_train $dumpdir/train-clean-360.json \
-		--clean_speech_valid $dumpdir/dev-clean.json \
-	  --rir_train $dumpdir/train.json \
-	  --rir_valid dumpdir/validation.json \
-		--sample_rate $sample_rate \
+		--clean_speech_train $dumpdir/clean/train-clean-360.json \
+		--clean_speech_valid $dumpdir/clean/dev-clean.json \
+	  --rir_train $dumpdir/rirs/train.json \
+	  --rir_valid dumpdir/rirs/validation.json \
+		--fs $sample_rate \
 		--exp_dir ${expdir}/ | tee logs/train_${tag}.log
 	cp logs/train_${tag}.log $expdir/train.log
